swagger

+++ src/main/java/yoot/yoedu_backend/dto/courseclass/CourseClassResponse.java
@@ -34,7 +34,7 @@ public class CourseClassResponse {
 
     private Integer maxStudents;
 
-    private Double tuitionFee;
+    private float tuitionFee;
 
     private ClassStatus status;
